fix: 자잘한 뷰 오류, api 오류 수정 && 내 추억 뷰 이미지 띄우기
