+++ Odya-iOS/Core/TravelJournal/TravelJournalDetail/View/TravelJournalDetailBottomSheet.swift
@@ -8,9 +8,9 @@
 import SwiftUI
 
 struct JournalDetailBottomSheet: View {
-    
+
   // MARK: Properties
-    
+
   @EnvironmentObject var bottomSheetVM: BottomSheetViewModel
 
   let title: String
@@ -37,9 +37,9 @@ struct JournalDetailBottomSheet: View {
     contents = travelJournal.dailyJournals
     mates = travelJournal.travelMates
   }
- 
-    // MARK: Body
-    
+
+  // MARK: Body
+
   var body: some View {
     ZStack {
       OffsettableScrollView { point in
@@ -55,15 +55,32 @@ struct JournalDetailBottomSheet: View {
         VStack(spacing: 24) {
           viewModeSelector
 
-            ForEach(self.contents, id: \.id) { dailyJournal in
-              let dayN: Int = Int(dailyJournal.travelDate.timeIntervalSince(self.startDate))
-            DailyJournalView(
-              dayN: dayN, dailyJournal: dailyJournal, isFeedType: $isFeedType,
-              isAllExpanded: $isAllExpanded)
+          ForEach(self.contents, id: \.id) { dailyJournal in
+            let dayN: Int = Int(dailyJournal.travelDate.timeIntervalSince(self.startDate) / 86400)
+            HStack(alignment: .top, spacing: 16) {
+              VStack(spacing: 12) {
+                Text("Day\(dayN + 1)")
+                  .b1Style()
+                  .foregroundColor(.odya.label.normal)
+                  .frame(height: 12)
+                  .padding(.bottom, 7)
+                SplarkleIcon()
+                if dailyJournal != self.contents.last {
+                  Rectangle()
+                    .frame(width: 4)
+                    .foregroundColor(Color.odya.elevation.elev6)
+                }
+              }.frame(width: 50)
+
+              DailyJournalView(
+                dailyJournal: dailyJournal, isFeedType: $isFeedType,
+                isAllExpanded: $isAllExpanded)
+            }
           }
 
         }
         .padding(.top, 25)
+        .padding(.bottom, 200)
         .padding(.horizontal, GridLayout.side)
         .background(Color.odya.elevation.elev1)
         .clipShape(RoundedEdgeShape(edgeType: .top, cornerRadius: Radius.large))
@@ -84,55 +101,55 @@ struct JournalDetailBottomSheet: View {
     .clipShape(RoundedEdgeShape(edgeType: .top, cornerRadius: Radius.medium))
     .ignoresSafeArea(edges: [.bottom])
     .sheet(isPresented: $showTravelMateList) {
-        TravelMatesView(mates: mates)
-            .presentationDetents([.medium])
+      TravelMatesView(mates: mates)
+        .presentationDetents([.medium])
     }
 
   }
 
-    // MARK: Journal Info
-    private var journalInfo: some View {
-        VStack(alignment: .leading) {
-            // 여행 제목
-            Text(title)
-                .h5Style()
-                .foregroundColor(.odya.label.normal)
-                .lineLimit(bottomSheetVM.isSheetOn ? 1 : nil)
-                .padding(.vertical)
-            
-            HStack {
-                Button(action: {}) {
-                    HStack(spacing: 0) {
-                        let displayedMates = mates.filter({ $0.isRegistered }).prefix(2)
-                        ForEach(Array(displayedMates.enumerated()), id: \.element.id) { index, mate in
-                            if let url = mate.profileUrl {
-                                ProfileImageView(profileUrl: url, size: .S)
-                                    .offset(x: index == 1 ? -8 : 0)
-                            }
-                        }
-                        
-                        if mates.count > 2 {
-                            Text("더보기")
-                                .detail2Style()
-                                .foregroundColor(.odya.label.normal)
-                                .padding(.vertical, 10)
-                                .padding(.horizontal, 12)
-                                .offset(x: -8)
-                        }
-                    }
-                }
-                
-                Spacer()
-                
-                // 여행 기간
-                Text("\(startDateString) ~ \(endDateString)")
-                    .detail2Style()
-                    .foregroundColor(.odya.label.assistive)
-            }
-        }
-    }
+  // MARK: Journal Info
+  private var journalInfo: some View {
+    VStack(alignment: .leading) {
+      // 여행 제목
+      Text(title)
+        .h5Style()
+        .foregroundColor(.odya.label.normal)
+        .lineLimit(bottomSheetVM.isSheetOn ? 1 : nil)
+        .padding(.vertical)
 
-    // MARK: View Mode Selector
+      HStack {
+        Button(action: {}) {
+          HStack(spacing: 0) {
+            let displayedMates = mates.filter({ $0.isRegistered }).prefix(2)
+            ForEach(Array(displayedMates.enumerated()), id: \.element.id) { index, mate in
+              if let url = mate.profileUrl {
+                ProfileImageView(profileUrl: url, size: .S)
+                  .offset(x: index == 1 ? -8 : 0)
+              }
+            }
+
+            if mates.count > 2 {
+              Text("더보기")
+                .detail2Style()
+                .foregroundColor(.odya.label.normal)
+                .padding(.vertical, 10)
+                .padding(.horizontal, 12)
+                .offset(x: -8)
+            }
+          }
+        }
+
+        Spacer()
+
+        // 여행 기간
+        Text("\(startDateString) ~ \(endDateString)")
+          .detail2Style()
+          .foregroundColor(.odya.label.assistive)
+      }
+    }
+  }
+
+  // MARK: View Mode Selector
   private var viewModeSelector: some View {
     HStack {
       Button(action: {
@@ -158,7 +175,7 @@ struct JournalDetailBottomSheet: View {
 
 struct JournalDetailBottomSheet_Previews: PreviewProvider {
   static var previews: some View {
-      TravelJournalDetailView(journalId: 1)
+    TravelJournalDetailView(journalId: 1)
     //        JournalDetailBottomSheet(isSheetOn: .constant(.journalDetailOn), travelJournal: TravelJournalData.getDummy())
   }
 }
